+++ core/src/main/java/com/taobao/arthas/core/shell/session/Session.java
@@ -92,7 +92,7 @@ public interface Session {
      *
      * @return java pid
      */
-    int getPid();
+    long getPid();
 
     /**
      * Get all registered command resolvers
